add is_vote_transaction
