+++ core/src/structures/produced_block.rs
@@ -1,4 +1,4 @@
-use solana_sdk::feature_set::full_inflation::mainnet::certusone::vote;
+use log::info;
 use solana_sdk::{
     borsh0_10::try_from_slice_unchecked,
     commitment_config::CommitmentConfig,
@@ -6,6 +6,8 @@ use solana_sdk::{
     slot_history::Slot,
     transaction::TransactionError,
 };
+use solana_sdk::program_utils::limited_deserialize;
+use solana_sdk::vote::instruction::VoteInstruction;
 use solana_transaction_status::{
     option_serializer::OptionSerializer, Reward, RewardType, UiConfirmedBlock,
     UiTransactionStatusMeta,
@@ -40,140 +42,6 @@ pub struct ProducedBlock {
 }
 
 impl ProducedBlock {
-    pub fn from_ui_block(
-        block: UiConfirmedBlock,
-        slot: Slot,
-        commitment_config: CommitmentConfig,
-    ) -> Self {
-        let block_height = block.block_height.unwrap_or_default();
-        let txs = block.transactions.unwrap_or_default();
-
-        let blockhash = block.blockhash;
-        let previous_blockhash = block.previous_blockhash;
-        let parent_slot = block.parent_slot;
-        let rewards = block.rewards.clone();
-
-        let txs = txs
-            .into_iter()
-            .filter_map(|tx| {
-                let Some(UiTransactionStatusMeta {
-                    err,
-                    compute_units_consumed,
-                    ..
-                }) = tx.meta
-                else {
-                    // ignoring transaction
-                    log::info!("Tx with no meta");
-                    return None;
-                };
-
-                let Some(tx) = tx.transaction.decode() else {
-                    // ignoring transaction
-                    log::info!("Tx could not be decoded");
-                    return None;
-                };
-
-                let signature = tx.signatures[0].to_string();
-                let cu_consumed = match compute_units_consumed {
-                    OptionSerializer::Some(cu_consumed) => Some(cu_consumed),
-                    _ => None,
-                };
-
-                let legacy_compute_budget = tx.message.instructions().iter().find_map(|i| {
-                    if i.program_id(tx.message.static_account_keys())
-                        .eq(&compute_budget::id())
-                    {
-                        if let Ok(ComputeBudgetInstruction::RequestUnitsDeprecated {
-                            units,
-                            additional_fee,
-                        }) = try_from_slice_unchecked(i.data.as_slice())
-                        {
-                            return Some((units, additional_fee));
-                        }
-                    }
-                    None
-                });
-
-                let mut cu_requested = tx.message.instructions().iter().find_map(|i| {
-                    if i.program_id(tx.message.static_account_keys())
-                        .eq(&compute_budget::id())
-                    {
-                        if let Ok(ComputeBudgetInstruction::SetComputeUnitLimit(limit)) =
-                            try_from_slice_unchecked(i.data.as_slice())
-                        {
-                            return Some(limit);
-                        }
-                    }
-                    None
-                });
-
-                let mut prioritization_fees = tx.message.instructions().iter().find_map(|i| {
-                    if i.program_id(tx.message.static_account_keys())
-                        .eq(&compute_budget::id())
-                    {
-                        if let Ok(ComputeBudgetInstruction::SetComputeUnitPrice(price)) =
-                            try_from_slice_unchecked(i.data.as_slice())
-                        {
-                            return Some(price);
-                        }
-                    }
-
-                    None
-                });
-
-                if let Some((units, additional_fee)) = legacy_compute_budget {
-                    cu_requested = Some(units);
-                    if additional_fee > 0 {
-                        prioritization_fees = Some(calc_prioritization_fees(units, additional_fee))
-                    }
-                };
-
-                let blockhash = tx.message.recent_blockhash().to_string();
-                let message = BinaryEncoding::Base64.encode(tx.message.serialize());
-
-                let is_vote_transaction = tx.message.instructions().iter().any(|i| {
-                    i.program_id(tx.message.static_account_keys())
-                        .eq(&vote::id())
-                });
-
-                Some(TransactionInfo {
-                    signature,
-                    is_vote: is_vote_transaction,
-                    err,
-                    cu_requested,
-                    prioritization_fees,
-                    cu_consumed,
-                    recent_blockhash: blockhash,
-                    message,
-                })
-            })
-            .collect();
-
-        let leader_id = if let Some(rewards) = block.rewards {
-            rewards
-                .iter()
-                .find(|reward| Some(RewardType::Fee) == reward.reward_type)
-                .map(|leader_reward| leader_reward.pubkey.clone())
-        } else {
-            None
-        };
-
-        let block_time = block.block_time.unwrap_or(0) as u64;
-
-        ProducedBlock {
-            transactions: txs,
-            block_height,
-            leader_id,
-            blockhash,
-            previous_blockhash,
-            parent_slot,
-            block_time,
-            slot,
-            commitment_config,
-            rewards,
-        }
-    }
-
     /// moving commitment level to finalized
     pub fn to_finalized_block(&self) -> Self {
         ProducedBlock {
@@ -181,19 +49,4 @@ impl ProducedBlock {
             ..self.clone()
         }
     }
-}
-
-#[inline]
-fn calc_prioritization_fees(units: u32, additional_fee: u32) -> u64 {
-    (units as u64 * 1000) / additional_fee as u64
-}
-
-#[test]
-fn overflow_u32() {
-    // value high enough to overflow u32 if multiplied by 1000
-    let units: u32 = 4_000_000_000;
-    let additional_fee: u32 = 100;
-    let prioritization_fees: u64 = calc_prioritization_fees(units, additional_fee);
-
-    assert_eq!(40_000_000_000, prioritization_fees);
 }
